Should sort trips with lowest price first
The test makes two requests in order to get a list of itineraries for a 
trip between Berlin and Prague with train. The first request is needed in 
order to get the search id and to save all the parameters in the search.
The second request specifies the sorting and return the list of itineraries. 
The verification is done on the service level, since the front-end will 
display the items in exactly the order they are presented.